Hide widgets with no component
