+++ web/src/beta/lib/core/Crust/Widgets/Widget/builtin/index.ts
@@ -1,17 +1,18 @@
 import { merge } from "lodash-es";
 
+import {
+  BUTTON_BUILTIN_WIDGET_ID,
+  NAVIGATOR_BUILTIN_WIDGET_ID,
+} from "@reearth/services/api/widgetsApi/utils";
+
 import Button from "./Button";
 import Navigator from "./Navigator";
-import Timeline from "./Timeline";
+// import Timeline from "./Timeline";
 import { Component, unsafeBuiltinWidgets, type UnsafeBuiltinWidgets } from "./unsafeWidgets";
-
-export const BUTTON_BUILTIN_WIDGET_ID = "reearth/button";
-export const TIMELINE_BUILTIN_WIDGET_ID = "reearth/timeline";
-export const NAVIGATOR_BUILTIN_WIDGET_ID = "reearth/navigator";
 
 export type ReEarthBuiltinWidgets<T = unknown> = Record<
   | typeof BUTTON_BUILTIN_WIDGET_ID
-  | typeof TIMELINE_BUILTIN_WIDGET_ID
+  // | typeof TIMELINE_BUILTIN_WIDGET_ID
   | typeof NAVIGATOR_BUILTIN_WIDGET_ID,
   T
 >;
@@ -20,9 +21,9 @@ export type BuiltinWidgets<T = unknown> = ReEarthBuiltinWidgets<T> & UnsafeBuilt
 
 const REEARTH_BUILTIN_WIDGET_OPTIONS: BuiltinWidgets<{ animation?: boolean }> = {
   [BUTTON_BUILTIN_WIDGET_ID]: {},
-  [TIMELINE_BUILTIN_WIDGET_ID]: {
-    animation: true,
-  },
+  // [TIMELINE_BUILTIN_WIDGET_ID]: {
+  //   animation: true,
+  // },
   [NAVIGATOR_BUILTIN_WIDGET_ID]: {
     animation: true,
   },
@@ -37,7 +38,7 @@ Object.keys(unsafeBuiltinWidgets ?? {}).map(uw => {
 
 const reearthBuiltin: BuiltinWidgets<Component> = {
   [BUTTON_BUILTIN_WIDGET_ID]: Button,
-  [TIMELINE_BUILTIN_WIDGET_ID]: Timeline,
+  // [TIMELINE_BUILTIN_WIDGET_ID]: Timeline,
   [NAVIGATOR_BUILTIN_WIDGET_ID]: Navigator,
 };
