+++ web/src/beta/lib/core/StoryPanel/Block/builtin/index.ts
@@ -1,23 +1,18 @@
 import { merge } from "lodash-es";
 
+import {
+  IMAGE_BUILTIN_STORY_BLOCK_ID,
+  TEXT_BUILTIN_STORY_BLOCK_ID,
+  TITLE_BUILTIN_STORY_BLOCK_ID,
+  VIDEO_BUILTIN_STORY_BLOCK_ID,
+} from "@reearth/services/api/storytellingApi/blocks";
+
 import { Component } from "..";
 
 import ImageBlock from "./Image";
 import TextBlock from "./Text";
 import TitleBlock from "./Title";
 import VideoBlock from "./Video";
-
-export const TITLE_BUILTIN_STORY_BLOCK_ID = "reearth/titleStoryBlock"; // pseudo storyblock
-
-export const IMAGE_BUILTIN_STORY_BLOCK_ID = "reearth/imageStoryBlock";
-export const TEXT_BUILTIN_STORY_BLOCK_ID = "reearth/textStoryBlock";
-export const VIDEO_BUILTIN_STORY_BLOCK_ID = "reearth/videoStoryBlock";
-
-export const AVAILABLE_STORY_BLOCK_IDS = [
-  IMAGE_BUILTIN_STORY_BLOCK_ID,
-  TEXT_BUILTIN_STORY_BLOCK_ID,
-  VIDEO_BUILTIN_STORY_BLOCK_ID,
-];
 
 export type ReEarthBuiltinStoryBlocks<T = unknown> = Record<
   | typeof TITLE_BUILTIN_STORY_BLOCK_ID
