+++ web/src/services/api/storytellingApi/blocks.ts
@@ -1,7 +1,6 @@
 import { useMutation, useQuery } from "@apollo/client";
 import { useCallback, useMemo } from "react";
 
-import { AVAILABLE_STORY_BLOCK_IDS } from "@reearth/beta/lib/core/StoryPanel/Block/builtin";
 import { MutationReturn } from "@reearth/services/api/types";
 import {
   CreateStoryBlockInput,
@@ -27,6 +26,18 @@ import { useNotification } from "@reearth/services/state";
 
 import { Item, convert } from "../propertyApi/utils";
 import { SceneQueryProps } from "../sceneApi";
+
+export const TITLE_BUILTIN_STORY_BLOCK_ID = "reearth/titleStoryBlock"; // pseudo storyblock
+
+export const IMAGE_BUILTIN_STORY_BLOCK_ID = "reearth/imageStoryBlock";
+export const TEXT_BUILTIN_STORY_BLOCK_ID = "reearth/textStoryBlock";
+export const VIDEO_BUILTIN_STORY_BLOCK_ID = "reearth/videoStoryBlock";
+
+export const AVAILABLE_STORY_BLOCK_IDS = [
+  IMAGE_BUILTIN_STORY_BLOCK_ID,
+  TEXT_BUILTIN_STORY_BLOCK_ID,
+  VIDEO_BUILTIN_STORY_BLOCK_ID,
+];
 
 export type StoryBlockQueryProps = SceneQueryProps;
